[Font switch] : (-) RoboKoz => (+) Fondamento
*RoboKoz* is distributed by certain fonts websites as public domain and cc-0 ; other website provide the same font with a README from the author stating : 
    "No payment is required for the use of this font. It's free!
    Commercial or personal use? It doesn't matter, it's free."
    (ref:  http://dafont.me/view_file/175/UmVhZF9NZS50eHQ=/read_me.txt )
This statements unfortunately doesn't explicit 'public domain or CC-0' nor open-source , and so, it's difficult to know if derivations are accepted... RoboKoz was needing a refactor to add accents, and developp it. It's too dangerous to do development upon a font we can't be 100% sure to be open-source or public-domain. 

*Fondamento* , on the other hand, is an OFL / SIL Open Font License. I add it as an alternative to 'Robokoz'. It also already has many accents built-in.
Link: https://www.google.com/fonts/specimen/Fondamento

(Note : no problem with keeping RoboKoz in Git tree , file is unchanged, and correctly attributed all along so it respect the author and his conditions.)