refactor of the string [warning witch property] in cover also
This text is not anymore a graphical text part of the artwork/drawing, 
but a text you can translate. It use the font Fondamento, part of the official
font since yesterday. This commit affect every language. ( already fixed in FR )

Note : no hurry ; the english text everywhere is like it was before anyway.